# Conflicts:
#	src/main/java/dcsc/mvc/util/Link.java

+++ src/main/java/dcsc/mvc/service/classes/ClassesServiceImpl.java
@@ -1,23 +1,27 @@
 package dcsc.mvc.service.classes;
 
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.util.Date;
 import java.util.List;
 
 import javax.transaction.Transactional;
 
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import com.querydsl.core.BooleanBuilder;
-
 import dcsc.mvc.domain.classes.ClassCategory;
+import dcsc.mvc.domain.classes.ClassImage;
 import dcsc.mvc.domain.classes.ClassSchedule;
+import dcsc.mvc.domain.classes.ClassState;
 import dcsc.mvc.domain.classes.Classes;
-import dcsc.mvc.domain.classes.QClasses;
 import dcsc.mvc.domain.classes.Search;
 import dcsc.mvc.repository.classes.ClassCategoryRepository;
 import dcsc.mvc.repository.classes.ClassImageRepository;
 import dcsc.mvc.repository.classes.ClassScheduleRepository;
 import dcsc.mvc.repository.classes.ClassesRepository;
 import lombok.RequiredArgsConstructor;
+import oracle.sql.TIMESTAMP;
 
 @Service
 @Transactional
@@ -33,10 +37,20 @@ public class ClassesServiceImpl implements ClassesService {
 	 * @param Class(선생님ID, 클래스명, 클래스 소개글, 클래스 금액, 카테고리);
 	 * */
 	@Override
-	public void insert(Classes classes) {
+	public void insert(Classes classes, ClassImage mainImage, ClassImage[] subImages) {
+		classes.setClassState(new ClassState(1L, null));
 		Classes dbClass = classesRepository.save(classes);
+		System.out.println(dbClass.getClassId());
 		
-		if(dbClass == null) throw new RuntimeException("클래스 등록을 실패했습니다.");
+		mainImage.setClasses(dbClass);
+		classImageRepository.save(mainImage);
+		
+		for(ClassImage subImage : subImages) {
+			if(subImage != null) {
+				subImage.setClasses(classes);
+				classImageRepository.save(subImage);
+			}
+		}
 	}
 
 	/**
@@ -61,7 +75,50 @@ public class ClassesServiceImpl implements ClassesService {
 	 * */
 	@Override
 	public void delete(Long classId) {
-		classesRepository.deleteById(classId);
+		Classes dbClass = classesRepository.findById(classId).orElse(null);
+		
+		if(dbClass == null) throw new RuntimeException("존재하지 않는 클래스입니다.");
+		if(dbClass.getClassState().getStateId() == 3) throw new RuntimeException("공개 클래스는 삭제할 수 없습니다.");
+		
+		dbClass.setClassState(new ClassState(5L, null));
+
+	}
+
+	/**
+	 * 클래스 공개 수락
+	 * @param Long classId;
+	 * */
+	@Override
+	public void updateStateOpen(Long classId) {
+		Classes dbClass = classesRepository.findById(classId).orElse(null);
+		
+		if(dbClass == null) throw new RuntimeException("존재하지 않는 클래스입니다.");
+		if(dbClass.getClassState().getStateId() != 2) throw new RuntimeException("공개 신청된 클래스가 아닙니다.");
+		
+		dbClass.setClassOpenDate(LocalDateTime.now());
+		dbClass.setClassState(new ClassState(3L, null));
+	}
+
+	/**
+	 * 클래스 상태 변경
+	 * @param Long classId, Long stateId(변경한 stateId);
+	 * */
+	@Override
+	public void updateState(Long classId, Long stateId) {
+		Classes dbClass = classesRepository.findById(classId).orElse(null);
+		
+		if(stateId == 1 && dbClass.getClassState().getStateId() != 2) throw new RuntimeException("공개 신청된 클래스가 아닙니다.");
+		else if(stateId == 2 && dbClass.getClassState().getStateId() != 1) throw new RuntimeException("공개 신청할 수 있는 클래스가 아닙니다.");
+		else if(stateId == 3 && dbClass.getClassState().getStateId() != 4) throw new RuntimeException("비공개된 클래스가 아닙니다.");
+		else if(stateId == 4 && dbClass.getClassState().getStateId() != 3) throw new RuntimeException("공개된 클래스가 아닙니다.");
+		else if(stateId == 4) {
+			for(ClassSchedule schedule : selectScheduleByClassId(classId)) {
+				if(schedule.getScheduleFinished().equals("T")) continue;
+				if(schedule.getTotalSeat() == schedule.getLeftSeat()) continue;
+				throw new RuntimeException("수강 신청된 일정이 있어 클래스를 비공개할 수 없습니다.");
+			}
+		}
+		dbClass.setClassState(new ClassState(stateId, null));
 	}
 
 	/**
@@ -82,12 +139,18 @@ public class ClassesServiceImpl implements ClassesService {
 	 * */
 	@Override
 	public List<Classes> selectByFilter(Search search) {
-		BooleanBuilder booleanBuilder = new BooleanBuilder();
-		
-		QClasses classes = QClasses.classes;
-		
-		if(search.getRegion() != null) booleanBuilder.and(classes.placeRegion.regionId.eq(search.getRegion()));
-		
+//		BooleanBuilder booleanBuilder = new BooleanBuilder();
+//		
+//		QClasses classes = QClasses.classes;
+//		
+//		if(search.getRegion() != null) booleanBuilder.and(classes.teacher.place.placeRegion.regionId.eq(search.getRegion()));
+//		if(search.getCategory() != null) booleanBuilder.and(classes.classCategory.categoryId.eq(search.getCategory()));
+//		
+//		LocalDate from = LocalDate.now();
+//		LocalDate to 
+////		if(search.getNewClass() == "T") booleanBuilder.and(classes.classSchedule.);
+//
+//		
 		return null;
 	}
 
@@ -137,14 +200,18 @@ public class ClassesServiceImpl implements ClassesService {
 		if(dbSchedule == null) throw new RuntimeException("해당 일정이 존재하지 않습니다");
 
 		dbSchedule.setScheduleDate(schedule.getScheduleDate());
-		dbSchedule.setStartTime(schedule.getStartTime());
-		dbSchedule.setEndTime(schedule.getEndTime());
-		dbSchedule.setTotalSeat(schedule.getTotalSeat());
+		
+		if(schedule.getTotalSeat() - dbSchedule.getLeftSeat() == 0) {
+			dbSchedule.setStartTime(schedule.getStartTime());
+			dbSchedule.setEndTime(schedule.getEndTime());
+		}
 		
 		int leftSeat = dbSchedule.getLeftSeat() - (dbSchedule.getTotalSeat() - schedule.getTotalSeat());
-		if(leftSeat < 0) throw new RuntimeException("변경할 예약 가능 인원이 예약한 인원보다 적습니다.");
+		if(leftSeat < 0) throw new RuntimeException("변경할 수강 가능 인원이 수강 신청자의 수보다 적어 변경할 수 없습니다.");
+		else if(schedule.getTotalSeat() < schedule.getLeftSeat()) throw new RuntimeException("수강 가능 인원은 남은 수강 인원보다 작을 수 없습니다.");
 		
-		dbSchedule.setLeftSeat(leftSeat);
+		dbSchedule.setLeftSeat(schedule.getLeftSeat());
+		dbSchedule.setTotalSeat(schedule.getTotalSeat());
 	}
 
 	/**
@@ -153,6 +220,10 @@ public class ClassesServiceImpl implements ClassesService {
 	 * */
 	@Override
 	public void deleteSchedule(Long scheduleId) {
+		ClassSchedule dbSchedule = classScheduleRepository.findById(scheduleId).orElse(null);
+		if(dbSchedule.getTotalSeat() != dbSchedule.getLeftSeat()) {
+			throw new RuntimeException("수강 신청자가 있을 경우 일정을 삭제할 수 없습니다.");
+		}
 		classScheduleRepository.deleteById(scheduleId);
 	}
